phan quyen cong trinh

+++ src/views/projects/ListProjects.js
@@ -11,7 +11,7 @@ import { DeleteOutline } from '@mui/icons-material';
 import MainCard from 'ui-component/cards/MainCard';
 
 import config from '../../config';
-import { apiGet, apiDelete } from '../../utils/formatUtils';
+import { apiGet, apiDelete, getRoles } from '../../utils/formatUtils';
 
 const LIST_PROJECTS = `${config.API_URL}/projects`;
 
@@ -19,9 +19,45 @@ export default function ListProjects() {
   const [open, setOpen] = useState(false);
   const [data, setData] = useState([]);
 
+  const [dataRoles, setDataRoles] = useState([]);
+  const [permissionAdd, setPermissionAdd] = useState(false);
+  const [permissionUpdate, setPermissionUpdate] = useState(false);
+  const [permissionDelete, setPermissionDelete] = useState(false);
+
   useEffect(() => {
+    loadListRoles();
     loadListProjects();
   }, []);
+
+  useEffect(() => {
+    if (dataRoles.length > 0) {
+      const filteredAdd = dataRoles.filter((role_add) => role_add.function_id === '667463d04bede188dfb46d7b');
+      const filteredUpdate = dataRoles.filter((role_update) => role_update.function_id === '667463d04bede188dfb46d7c');
+      const filteredDelete = dataRoles.filter((role_delete) => role_delete.function_id === '667463d04bede188dfb46c7c');
+      if (filteredAdd.length > 0) {
+        setPermissionAdd(true);
+      } else {
+        setPermissionAdd(false);
+      }
+
+      if (filteredUpdate.length > 0) {
+        setPermissionUpdate(true);
+      } else {
+        setPermissionUpdate(false);
+      }
+
+      if (filteredDelete.length > 0) {
+        setPermissionDelete(true);
+      } else {
+        setPermissionDelete(false);
+      }
+    }
+  }, [dataRoles]);
+
+  const loadListRoles = async () => {
+    const result = await getRoles();
+    setDataRoles(result.data);
+  };
 
   const loadListProjects = async () => {
     const result = await apiGet(`${LIST_PROJECTS}`);
@@ -44,30 +80,38 @@ export default function ListProjects() {
 
   const columns = [{ field: 'name', headerName: 'Tên trạng thái', width: 300 }];
 
-  columns.push({
-    field: 'action',
-    headerName: 'Hành động',
-    width: 120,
-    renderCell: (params) => {
-      return (
-        <>
-          <Link to={'/dashboard/projects/update-projects/' + params.row._id}>
-            <IconEdit />
-          </Link>
-          <DeleteOutline style={{ cursor: 'pointer', color: '#ff6666' }} onClick={() => handleDelete(params.row._id)} />
-        </>
-      );
-    }
-  });
+  if (permissionUpdate || permissionDelete) {
+    columns.push({
+      field: 'action',
+      headerName: 'Hành động',
+      width: 120,
+      renderCell: (params) => {
+        return (
+          <>
+            {permissionUpdate && (
+              <Link to={'/dashboard/projects/update-projects/' + params.row._id}>
+                <IconEdit />
+              </Link>
+            )}
+            {permissionDelete && (
+              <DeleteOutline style={{ cursor: 'pointer', color: '#ff6666' }} onClick={() => handleDelete(params.row._id)} />
+            )}
+          </>
+        );
+      }
+    });
+  }
 
   return (
     <>
       <MainCard
         title="Danh sách"
         secondary={
-          <Button variant="contained" component={Link} to="/dashboard/projects/add-projects">
-            Thêm mới
-          </Button>
+          permissionAdd && (
+            <Button variant="contained" component={Link} to="/dashboard/projects/add-projects">
+              Thêm mới
+            </Button>
+          )
         }
       >
         {data.length !== 0 ? (
